+++ opentelemetry-otlp/src/exporter/http/logs.rs
@@ -36,6 +36,10 @@ impl LogExporter for OtlpHttpClient {
     fn set_resource(&mut self, resource: &opentelemetry_sdk::Resource) {
         self.resource = resource.into();
     }
+
+    fn requires_async_runtime(&self) -> bool {
+        self.requires_async_runtime()
+    }
 }
 
 /// Handles partial success returned by OTLP endpoints. We log the rejected log records,